starting to debug
